# contracts/openapi/core-api.v1.yaml
openapi: 3.0.3
info:
  title: FLIT Core API
  version: "1.0.0"
  description: >
    Contrato público de FLIT Core API (services/core-api). Cubre el carril de
    Seguridad/Autenticación (HU #10168 — login email/contraseña y emisión JWT 12h)
    y el módulo Administrador de Compañías (HU #10189 — index paginado y
    seguridad SuperAdmin).
servers:
  - url: https://api.flit.co
    description: Producción (vía Gateway YARP)
  - url: http://localhost:4002
    description: Local (Gateway)
...
paths:
...
  /api/v1/tramites:
    post:
      tags: [Trámites · Documentos]
      operationId: tramitesCreate
      summary: Crea un trámite y congela su snapshot documental (RF19)
      description: >
        Crea un trámite en estado draft y captura, en la misma transacción, un snapshot
        inmutable de la matriz documental resuelta vigente (precedencia Cliente > OT >
        Default). Si la generación del snapshot falla, toda la operación se revierte y no
        queda trámite creado (500). Requiere JWT autenticado (no SuperAdmin).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcedureInstanceRequest"
      responses:
        "201":
          description: Trámite creado con su snapshot documental.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcedureInstanceResponse"
        "401":
          description: No autenticado (sin token JWT).
        "404":
          description: El tipo de trámite o el cliente (tenant) no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Datos inválidos o número de referencia duplicado para el cliente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Falló la persistencia del trámite o del snapshot; operación revertida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/tramites/instances/{id}/finalize-draft:
    post:
      operationId: finalizeDraftProcedureInstance
      tags: [Tramites]
      summary: Finalizar borrador del trámite (HU #10349)
      description: >
        Marca un borrador como "finalizado": valida que la captura de datos esté completa (actores de la
        modalidad, documentos obligatorios y organismo de tránsito) PERO **no** exige validación de
        identidad ni FUR. La instancia permanece en estado `draft` y se sella `draftFinalizedAt`, quedando
        a la espera de la validación de identidad asíncrona del cliente (fase 2). Idempotente: si el
        borrador ya estaba finalizado, devuelve el resumen actual sin re-sellar. La radicación real
        (`submit`, Draft→Submitted) sigue exigiendo identidad + FUR + reglas OT.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: Identificador de la instancia (procedure_instance_id).
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Borrador finalizado; `draftFinalizedAt` sellado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcedureInstanceSummary"
        "400":
          description: Falta el header X-Tenant-Id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: La instancia no existe para el tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            La instancia no está en `draft` (`not_draft`) o faltan datos para finalizar
            (`actores_incompletos`, `documentos_incompletos`, `organismo_requerido`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    CreateProcedureInstanceRequest:
      type: object
      properties:
        procedureTypeId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        transitOfficeId:
          type: string
          format: uuid
          nullable: true
          description: Organismo de tránsito (nivel OT) — opcional, afecta la precedencia.
        referenceNumber:
          type: string
          maxLength: 30
          description: Número de referencia del trámite, único por cliente.
        createdByUserId:
          type: string
          format: uuid
          nullable: true
          description: Usuario creador; si se omite, se resuelve del JWT (sub).
      required: [procedureTypeId, tenantId, referenceNumber]
    ProcedureInstanceResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        referenceNumber: { type: string }
        status: { type: string, example: draft }
        snapshotAt: { type: string, format: date-time }
        documentCount: { type: integer, description: Documentos congelados en el snapshot. }
      required: [id, referenceNumber, status, snapshotAt, documentCount]
    ProcedureInstanceSummary:
      type: object
      description: Resumen de una instancia de trámite (devuelto por finalize-draft y submit).
      properties:
        id: { type: string, format: uuid }
        referenceNumber: { type: string }
        status: { type: string, example: draft }
        procedureTypeId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        submittedAt:
          type: string
          format: date-time
          nullable: true
          description: Fecha de radicación (Draft→Submitted). Null mientras es borrador.
        draftFinalizedAt:
          type: string
          format: date-time
          nullable: true
          description: >
            HU #10349 — fecha en que el borrador se finalizó (datos completos a la espera de la
            validación de identidad async). Null si el borrador no se ha finalizado.
      required: [id, referenceNumber, status, procedureTypeId, tenantId, createdAt]
...